pyproject.toml
name = "ai-qa-assistant"
version = "0.1.0"
description = "AI-powered Q&A assistant with tool use, FastAPI server, and enterprise-ready patterns"
requires-python = ">=3.10"
readme = "README.md"


[project.optional-dependencies]
all = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "openai>=1.40.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.1.0",
    "slowapi>=0.1.9",
    "redis>=5.0.7",
    "faiss-cpu>=1.8.0.post1",
    "numpy>=1.26.4",
    "typer>=0.12.3",
    "yfinance>=0.2.40",
    "tenacity>=9.0.0",
    "SQLAlchemy>=2.0.30"
]


[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000


[tool.pytest.ini_options]
addopts = "-q"